requirements.txt
Flask
requests
python-dotenv
flask_cors
flask-talisman
openai